+++ nion/utils/test/Binding_test.py
@@ -8,7 +8,6 @@ import unittest
 # local libraries
 from nion.utils import Binding
 from nion.utils import Event
-from nion.utils import ListModel
 
 
 class TupleModel:
@@ -34,40 +33,6 @@ class TestBindingClass(unittest.TestCase):
 
     def tearDown(self):
         pass
-
-    def test_should_be_able_to_get_items_from_list_binding(self):
-        list_model = ListModel.ListModel("items")
-        list_model.insert_item(0, "zero")
-        list_model.insert_item(1, "one")
-        list_model.insert_item(2, "two")
-        binding = Binding.ListBinding(list_model, "items")
-        items = binding.items
-        self.assertEqual(len(items), 3)
-        self.assertEqual(items[2], "two")
-        list_model.insert_item(0, "negative")
-        self.assertEqual(len(items), 4)
-        self.assertEqual(items[3], "two")
-
-    def test_inserting_and_removing_item_into_binding_notifies_target(self):
-        list_model = ListModel.ListModel("items")
-        binding = Binding.ListBinding(list_model, "items")
-        list_copy = list()
-
-        def inserter(value: str, index: int) -> None:
-            list_copy.insert(index, value)
-
-        def remover(index: int) -> None:
-            del list_copy[index]
-
-        binding.inserter = inserter
-        binding.remover = remover
-        list_model.insert_item(0, "zero")
-        list_model.insert_item(1, "one")
-        self.assertEqual(len(list_copy), 2)
-        self.assertEqual(list_copy[1], "one")
-        list_model.remove_item(0)
-        self.assertEqual(len(list_copy), 1)
-        self.assertEqual(list_copy[0], "one")
 
     def test_tuple_binding_pads_to_index_if_necessary(self):
         # this allows the source to more easily go from None to a partialy tuple None -> (3, None) -> (3, 4)